requirements.txt
Flask==3.0.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1
Werkzeug==3.0.3